Use shine/steam webview to display changelog
The old menu webview element is no longer available. So let's use
Shine's/Steam's one instead.

Also refactor the changelog to be a mod
module instead of a seperate mod to decrease code complexity.
