--- lua/ns2_beta/Client.lua
@@ -1,5 +1,0 @@
-do
-    if ModuleLoader then
-        ModuleLoader:LoadAllModules("Client")
-    end
-end
