--- lua/ns2_beta/Predict.lua
@@ -1,5 +1,0 @@
-do
-    if ModuleLoader then
-        ModuleLoader:LoadAllModules("Predict")
-    end
-end
