--- lua/ns2_beta/Server.lua
@@ -1,7 +1,0 @@
-do
-    if ModuleLoader then
-        ModuleLoader:LoadAllModules("Server")
-
-        Server.AddRestrictedFileHashes("lua/TBCD/modules/*.lua")
-    end
-end
