--- lua/ns2_beta/Shared.lua
@@ -1,7 +1,0 @@
-kBetaVersion = 5 -- 1 year month day versionofday
-
-do
-    if ModuleLoader then
-        ModuleLoader:LoadAllModules("Shared")
-    end
-end
